move sanitizeError() to the error service method
