+++ src/services/__tests__/errorService.test.js
@@ -13,7 +13,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
 
-import sanitizeError from 'util/sanitizeError';
+// import sanitizeError from 'util/sanitizeError';
+import errorService from 'services/errorService';
 import networkService from 'services/networkService';
 import store from 'store';
 jest.mock('store');
@@ -32,7 +33,7 @@ describe('sanitizeError', () => {
       code: 4001,
       message: 'MetaMask Tx Signature: User denied transaction signature'
     };
-    const res = await sanitizeError(error);
+    const res = await errorService.sanitizeError(error);
     expect(res).toBe(error.message);
   });
 
@@ -41,7 +42,7 @@ describe('sanitizeError', () => {
       code: -32000,
       message: 'header not found'
     };
-    const res = await sanitizeError(error);
+    const res = await errorService.sanitizeError(error);
     expect(res).toBe(null);
   });
 
@@ -50,7 +51,7 @@ describe('sanitizeError', () => {
       message: 'Insufficient funds. Needs 100000000000000000 more of 0x123 to cover payments and fees'
     };
     // this should fetch OMG of 18 decimals
-    const res = await sanitizeError(error)
+    const res = await errorService.sanitizeError(error)
     expect(res).toBe('Insufficient funds. Needs 0.1 more of OMG to cover payments and fees')
     // expect getToken to be called
     const expectedActions = [ {
@@ -66,7 +67,7 @@ describe('sanitizeError', () => {
 
   it('should return reason on EVM error if possible', async () => {
     const error = { message: exampleEVMError };
-    const res = await sanitizeError(error);
+    const res = await errorService.sanitizeError(error);
     expect(networkService.OmgUtil.ethErrorReason).toBeCalled();
     expect(networkService.OmgUtil.ethErrorReason).toBeCalledWith({
       web3: expect.anything(),
@@ -78,8 +79,8 @@ describe('sanitizeError', () => {
   it('should return sensible error on EVM error without reason', async () => {
     const error = { message: exampleEVMError };
     // run first mocked implementation
-    await sanitizeError(error);
-    const res = await sanitizeError(error);
+    await errorService.sanitizeError(error);
+    const res = await errorService.sanitizeError(error);
     expect(res).toBe('Transaction has been reverted by the EVM');
   });
 
@@ -88,13 +89,13 @@ describe('sanitizeError', () => {
       code: 1337,
       message: 'hello world'
     };
-    const res = await sanitizeError(error);
+    const res = await errorService.sanitizeError(error);
     expect(res).toBe(error.message);
   });
 
   it('should not crash on unusual error', async () => {
     const error = { foo: 'bar' };
-    const res = await sanitizeError(error);
+    const res = await errorService.sanitizeError(error);
     expect(res).toBe('Something went wrong');
   });
 });
