+++ src/services/errorService.js
@@ -14,7 +14,11 @@ See the License for the specific language governing permissions and
 limitations under the License. */
 
 import * as Sentry from '@sentry/browser';
-
+import networkService from 'services/networkService';
+import { getToken } from 'actions/tokenAction';
+import { logAmount } from 'util/amountConvert';
+import BigNumber from 'bignumber.js';
+import JSON5 from 'json5';
 import config from 'util/config';
 
 if (config.sentry) {
@@ -28,6 +32,48 @@ class ErrorService {
       Sentry.captureException(error);
     }
   }
+
+  async sanitizeError (error) {
+    // user sign rejection from metamask
+    if (error.code === 4001) {
+      return error.message;
+    }
+
+    // ignore metamask rpc header not found errors
+    if (error.code === -32000) {
+      return null;
+    }
+
+    // omg-js create UTXOs insufficient funds error
+    if (error.message && error.message.includes('Insufficient funds. Needs ')) {
+      const tokenAddress = error.message.split(' ').find(i => i.startsWith('0x'))
+      const token = await getToken(tokenAddress)
+      const tokenAmount = error.message.split(' ').find((i) => {
+        return i.match(/(?!0x)\d*/)[0] !== ""
+      })
+      const decimalAmount = logAmount(tokenAmount, token.decimals)
+      return `Insufficient funds. Needs ${decimalAmount.toString()} more of ${token.name} to cover payments and fees`
+    }
+
+    // try get reason from evm error message
+    const revertedMessage = 'Transaction has been reverted by the EVM:';
+    if (error.message && error.message.includes(revertedMessage)) {
+      try {
+        const errorTx = JSON5.parse(error.message.split(revertedMessage)[1]);
+        const reason = await networkService.OmgUtil.ethErrorReason({
+          web3: networkService.web3,
+          hash: errorTx.transactionHash
+        });
+        return reason ? reason : 'Transaction has been reverted by the EVM';
+      } catch (error) {
+        return 'Transaction has been reverted by the EVM';
+      }
+    }
+
+    // default to original error message
+    return error.message || 'Something went wrong';
+  }
+
 }
 
 const errorService = new ErrorService();
